# static/openapi.yaml
openapi: 3.0.3
info:
  version: 0.0.1
  title: Microservice - Product API document
  description: |-
    This is the API document of Product Microservice based on the OpenAPI 3.0 specification
      - Product
      - Cart
      - Order
servers:
  - url: https://13.58.156.142:5000
paths:
...
  /products/search:
    get:
      tags:
        - products
      summary: Search products
      description: Search products
      operationId: productsSearch
      parameters:
        - name: category
          in: query
          description: Category
          required: false
          schema:
            type: string
        - name: discount
          in: query
          description: Discount
          required: false
          schema:
            type: number
            format: float
        - name: stock
          in: query
          description: Stock
          required: false
          schema:
            type: integer
        - name: ordering
          in: query
          description: Ordering
          required: false
          schema:
            type: string
            default: asc
            enum:
              - asc
              - desc
        - name: page
          in: query
          description: Page
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductsGetResp'
...
components:
  schemas:
    ProductsGetResp:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: float
          example: 11.11
        stock:
          type: integer
          example: 10
        image:
          type: string
          example: "https://image.png"
        category:
          type: string
        discount:
          type: number
          format: float
          example: 11.11
        provider_id:
          type: integer
          example: 1
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      xml:
        name: products
...